// src/types/fieldTypes.ts
export type Field = {
  id: string;
  coordinates: { x: number; y: number };
  type: string;
  owner: string;
};